% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_batch.R
\name{mt_batch}
\alias{mt_batch}
\title{Correct for batch-effects in \code{mmt} objects based on variable/sample metadata.}
\usage{
mt_batch(data, ...)
}
\arguments{
\item{mmt}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}

\item{batch}{(\emph{required}) Variable in metadata to perform batch-correct on.}

\item{method}{(\emph{optional}) Specify the method for batch-correction. One of:
\itemize{
   \item \code{"ComBat"}: (\emph{default}) See \link[sva]{ComBat} for details and additional parameter settings.
}}

\item{...}{Additional parameters for the specified method.}
}
\value{
A modifed mt object
}
\description{
Correct for batch-effects in \code{mmt} objects based on variable/sample metadata.
}
\details{
The function \code{\link{mt_batch}} tries to eliminate batch-effects.
}
\examples{

\dontrun{
# EXAMPLE GOES HERE
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
