% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_diffexprs.R
\name{mt_diffexprs}
\alias{mt_diffexprs}
\title{mt_diffexprs}
\usage{
mt_diffexprs(obj, group, intercept = NULL, row_labels = NULL,
  order_var_by = "variance", row_show = 10)
}
\arguments{
\item{obj}{(\emph{required}) An object of class 'mmt' (see \code{\link{mt_load}}) or 'ampvis2' (see \code{amp_load}).}

\item{group}{(\emph{required}) Character string defining the grouping variable to use for differential expression.}

\item{intercept}{(\emph{optional}) Specify the level of 'group' to be used as reference/intercept. Default is by alphabetical order.}

\item{row_labels}{(\emph{optional}) Specify columns of rowdata to print on heatmap.}

\item{order_var_by}{(\emph{optional}) How to order rows of heatmap. You can specify 1) "variance" or 2) one of the levels in \code{group}, which will order according to the absolute value of that level. (\emph{default}: "variance")}

\item{row_show}{(\emph{optional}) The number of rows to show.}
}
\value{
A list with 2 elements:

\itemize{
  \item \strong{heatmap} - A heatmap showing the log fold-change of each level in \code{group} relative to the intercept/baseline level specified in \code{intercept}. The \code{intercept} is therefore not possible to plot.
  \item \strong{DESeq} - A DESeqDataSet containing the results of the differential expression analysis. To extract results and do comparisons between groups use the \code{\link{mt_dumpDE}} function.
}
}
\description{
This function calculates differential expression by grouping variable supplied by the user.
}
\details{
Typically one would subset low-expression genes (use \code{\link{mt_subset}}) before performing the analysis, due to a low signal/noise ratio. This also drastically reduces computation times and improves stability of the parameter estimation.
}
\examples{
\dontrun{
# Load data.
data("example_mmt")

# Compute the statistics.
DE_mt <- mt_diffexprs(example_mmt,
  group      = "Type",
  row_labels = c("GeneID","product"),
  intercept  = "ANAMMOX")

# Extract results for given levels.
mt_res <- mt_dumpDE(DE_mt,nom = "ELECTRODE",denom = "SUSPENTION")

# Show the output.
mt_res$MAplot
mt_res$BOXplot
head(mt_res$Table)
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
