% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_dumpDE.R
\name{mt_dumpDE}
\alias{mt_dumpDE}
\title{Extract results from DE analysis done using \code{mt_diffexprs}.}
\usage{
mt_dumpDE(mmtDE,num,denom)
}
\arguments{
\item{mmtDE}{(\emph{required}) An object of class \code{mmtDE}. See \code{\link{mt_diffexprs}}.}

\item{num}{(\emph{required}) Level that should be in the numerator when doing the DE analysis.}

\item{denom}{(\emph{required}) Level that should be in the denominator (i.e. the reference) when doing the DE analysis. If set to "rest", the DE analysis is done as a one-vs-all test. See XXX for details.}

\item{sub_genes}{The genes to include in the MA and boxplot. \emph{default}: all genes)}

\item{signif}{The threshold for adjusted p-value to be considered significant. (\emph{default}: 0.01)}

\item{text_MAplot_size}{Size of text on the MA plot. (\emph{default}: 3)}

\item{ngenes_show}{The number of genes to show in the boxplot. (\emph{default}: 10)}

\item{order_by}{Order the boxplot and table by a statistic from the DE analysis. One of: (\emph{default}: "logFC")}

\item{table_full}{(\emph{logical}) Output entire results table without filtering by significance. (\emph{default}: FALSE)}

\item{annot_size}{The size of text in the MA plot. (\emph{default}: 5)
\itemize{
  \item logFC - the absolute log2 fold-change betweem \code{num} and \code{denom}.
  \item padj - the adjusted p-value of the DE analysis.
}}

\item{...}{Additional arguments parsed to the \code{\link{vis_boxplot}} function.}
}
\value{
A list with 3 elements:
\itemize{
  \item \strong{MAplot} - A MAplot (ggplot-object) with the gene expression plotted against the log2 fold-change.
  \item \strong{BOXplot} - A faceted boxplot showing the expression of the most interesting genes, grouped according to the DE analysis design.
  \item \strong{Table} - A table with the variable metadata and associated statistics from the DE analysis.
}
}
\description{
mt_dumpDE
}
\details{

}
\examples{

\dontrun{
# Load data.
data("example_mmt")

# Compute the statistics.
DE_mt <- mt_diffexprs(example_mmt,
  group      = "Type",
  row_labels = c("GeneID","product"),
  intercept  = "ANAMMOX")

# Extract results for given levels.
mt_res <- mt_dumpDE(DE_mt,num = "ELECTRODE",denom = "SUSPENTION")

# Show the output.
mt_res$MAplot
mt_res$BOXplot
head(mt_res$Table)
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
