% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_enrichKO.R
\name{mt_enrichKO}
\alias{mt_enrichKO}
\title{Perform enrichment analysis using KEGG annotation.}
\usage{
mt_enrich(mmt,GeneIDs,KOs,type)
}
\arguments{
\item{mmt}{(\emph{required}) An object of class \code{mmt}.}

\item{GeneIDs}{(\emph{required}) Character vector with gene IDs of interest.}

\item{type}{(\emph{required}) String refering to which KEGG annotation scheme to use. Choose from the following:
\itemize{
  \item \strong{pathway} - Manually curated pathway maps of known molecular interaction, reaction and relation networks.
  \item \strong{module} - Manually defined functional units.
  \item \strong{reaction} - Chemical reactions, mostly enzymatic reactions.
  \item \strong{reaction_class} - Classes of reactions based on the chemical structure transformation patterns of substrate-product pairs.
}}

\item{KOs}{(\emph{required}) String pointing to the column in mtgene containing KO identifiers.}

\item{alternative}{(\emph{optional}) String specifying the direction of the test. Choose from:
\itemize{
  \item \strong{greater} - test for enrichment in genes of interest.
  \item \strong{less} - test for depletion in genes of interest.
  \item \strong{two.sided} - perform a two-sided test for both in genes of interest.
}}

\item{show_p}{(\emph{optional}) numeric indicating the threshold p-value for plotting. Default: 1}

\item{size_p}{(\emph{optional}) numeric indicating the size of plotted p-values. Default: 1}

\item{unannotated}{(\emph{optional}) Should the data for unannotated genes be included in plot? Default: \code{TRUE}.}
}
\value{
A list with following:
\itemize{
  \item \strong{table} - A table containing the results from the enrichment analysis.
  \item \strong{plot} - A ggplot barplot showing reduction/enrichment of each entry in \code{type} in percent, relative to the percent found in the entire dataset.
}
}
\description{
mt_enrichKO
}
\details{

}
\examples{

\dontrun{
# TBD.
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
