% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{mt_gather}
\alias{mt_gather}
\title{Gather the mmt object into one dataframe, suitable for e.g. ggplot or custom analysis (internal function)}
\usage{
mt_gather(mmt, metavars = c("SampleID", "GeneID"))
}
\arguments{
\item{mmt}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}

\item{metavars}{(\emph{optional}) Columns from mtgene/mtmeta to keep. (\emph{default:} \code{NULL} (all columns))}
}
\value{
A data.table in long format. The gene count/expression is stored in the \code{Exprs} column.
}
\description{
Gather the mmt object into one dataframe, suitable for e.g. ggplot or custom analysis (internal function)
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
