% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{mt_GenomeFrequencyMatrix}
\alias{mt_GenomeFrequencyMatrix}
\title{Computes a by-genome frequency matrix of any attribute in the mtgene data (internal function)}
\usage{
mt_GenomeFrequencyMatrix(mmt, Genome, ID)
}
\arguments{
\item{mmt}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}

\item{Genome}{(\emph{required}) Column name in mtgene grouping into genomes. NA and empty fields are ignored.}

\item{ID}{(\emph{required}) Column name in mtgene identifying the attribute. NA and empty fields are ignored.}
}
\value{
A data.table
}
\description{
Computes a by-genome frequency matrix of any attribute in the mtgene data (internal function)
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
