% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_phi.R
\name{mt_phi}
\alias{mt_phi}
\title{Calculates the phi-statistic of all genes in \code{mmt} object.}
\usage{
mt_subset(data, ...)
}
\arguments{
\item{mmt}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}
}
\value{
A genes x genes matrix with phi-statistics.
}
\description{
Calculate the phi-statistic for all genes in mmt object.
}
\details{
Some details here.
}
\examples{

\dontrun{
# An example goes here.
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
