% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plotpairs.R
\name{mt_plotpairs}
\alias{mt_plotpairs}
\title{Plot multiple combinations of sample gene expressions in a pairs plot}
\usage{
mt_plotpairs(mmt, samples = NULL, label_by = NULL, textsize = 5,
  pointsize = 2)
}
\arguments{
\item{mmt}{(\emph{required}) A \code{mmt} list loaded with \code{\link{mt_load}}.}

\item{samples}{A vector of 3 or more sample names in \code{mmt} to plot on each axis. If NULL, the default, then all samples will be plotted. (\emph{Default: } \code{NULL})}

\item{label_by}{replace the SampleIDs plotted in the diagonal with one(!) column in the metadata.}

\item{textsize}{The text size of the axis titles.}

\item{pointsize}{The size of points in the plot.}
}
\value{
A ggplot2 object.
}
\description{
Plots the gene expression from multiple samples in a \code{mmt} object in a grid plot with all pairs of variables.
}
\examples{
\dontrun{
data("example_mmt")

# Plot all samples (might take some time)
mt_plotpairs(example_mmt)

# Plot replicates and relabel.
mt_plotpairs(example_mmt,samples = c("HQ180323_13","HQ180323_14"),label_by = "Replicate")
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
