% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stackbar.R
\name{mt_stackbar}
\alias{mt_stackbar}
\title{Make a stacked barplot of the sequencing statistics.}
\usage{
mt_stackbar(mmt,Raw,after_QC)
}
\arguments{
\item{mmt}{(\emph{required}) An object of class \code{mmt}.}

\item{group_by}{(\emph{optional}) A column name in the metadata to group samples by. (\emph{Default: } \code{NULL})}

\item{round_to}{(\emph{optional}) Round the counts to thousands ("K") or milions ("M"). (\emph{Default: } \code{"M"})}

\item{size_axis}{(\emph{optional}) Specify size of primary and secondary axis. (\emph{Default: } \code{10})}

\item{size_text}{(\emph{optional}) Specify size of text in plot. (\emph{Default: } \code{3})}

\item{detailed}{(\emph{optional}) Output a table with the data used for the plot.}
}
\value{
A ggplot
}
\description{
mt_stackbar
}
\details{
The \code{\link{mt_stackbar}} function only works on data generated
from the \href{https://github.com/TYMichaelsen/MetaT}{MetaT} pipeline and loaded
using the \code{\link{mt_loadMetaT}} function. It is meant for data quality assesment
by giving a quick overview of the proportions of data removed for each pre-processing
step, if this is available in the metadata.
}
\examples{

\dontrun{
data("example_mmt")

mt_stackbar(example_mmt,Raw = "Raw",after_QC = "after_QC",after_filter = "after_filter")
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
