% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stattab.R
\name{mt_stattab}
\alias{mt_stattab}
\title{mt_stattab}
\usage{
mt_stattab(mt, add_meta = NULL, textsize = 7)
}
\arguments{
\item{mt}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}

\item{add_meta}{(\emph{optional}) Add an aditional column of metadata to the table, for explanatory reasons.}

\item{textsize}{(\emph{optional}) Adjust the textsize of the table.}
}
\value{
A ggplot object.
}
\description{
This function makes a table of the 'mtstat' component.
}
\details{
The \code{\link{mt_stattab}} function takes a 'mt' object and generates a table of sequencing statistics based on the sequencing stats stored in the object, if such are available. Throws an error if data is missing.
}
\examples{
Some example.

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
