% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_subset.R
\name{mt_subset}
\alias{mt_subset}
\title{Subset mt objects based on sample metadata}
\usage{
mt_subset(data, ...)
}
\arguments{
\item{mt}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}

\item{sub_genes}{(\emph{optional}) A string specifying a logical row subset operation on the mtgene dataframe in the mt object parsed to \code{subset()}.}

\item{sub_samples}{(\emph{optional}) A string specifying a logical row subset operation on the mtmeta dataframe in the mt object parsed to \code{subset()}.}

\item{minreads}{Minimum number of reads pr. gene. Genes below this value will be removed. (\emph{default:} \code{0})}
}
\value{
A modifed mt object
}
\description{
Subsets the data in ampvis2 objects based on metadata and returns the subsetted object.
}
\details{
The subset is performed on the mtmeta/mtgene data by \code{subset()} and the whole object is then adjusted accordingly.
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
