% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{print.mmt}
\alias{print.mmt}
\title{Prints mmt object summary (internal function)}
\usage{
\method{print}{mmt}(obj)
}
\arguments{
\item{obj}{(\emph{required}) Data list as loaded with \code{\link{mt_load}}.}
}
\description{
Prints mmt object summary (internal function)
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
