% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{setroworder}
\alias{setroworder}
\title{The row-version of setcolorder in data.table (internal function)}
\usage{
setroworder(x, neworder)
}
\arguments{
\item{x}{A \code{data.table}.}

\item{neworder}{Numeric vector of the row ordering.}
}
\value{
A data.table
}
\description{
The row-version of setcolorder in data.table (internal function)
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
