% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{test_pathway}
\alias{test_pathway}
\title{performs a hypergeometric test on a set of genes (internal function)}
\usage{
test_pathway(q_genes, p_genes, u_genes, alternative = "greater")
}
\arguments{
\item{q_genes}{(\emph{required}) genes in query.}

\item{p_genes}{(\emph{required}) genes in pathway of interest.}

\item{u_genes}{(\emph{required}) genes in universe.}

\item{alternative}{(\emph{required}) test for enrichment ('greater'), depletion ('less') or both ('two.sided').}
}
\value{
A named vector.
}
\description{
performs a hypergeometric test on a set of genes (internal function)
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
