% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_boxplot.R
\name{vis_boxplot}
\alias{vis_boxplot}
\title{Boxplot}
\usage{
vis_boxplot(obj,group_by)
}
\arguments{
\item{obj}{(\emph{required}) A data object of class 'ampvis2' (see \code{\link[ampvis2]{amp_load}}) or class 'mmt' (see \code{\link{mt_load}}).}

\item{group_by}{(\emph{required}) Group the samples by a variable in the metadata.}

\item{boxgroup_by}{Group the samples within each boxplot by a variable in the metadata.}

\item{sort_by}{Sort the boxplots by \code{"median"}, \code{"mean"} or \code{"total"}. (\emph{default:} \code{"median"})}

\item{normalise}{Can take one of following arguments: (\emph{default:} \code{"libsize"})
\itemize{
   \item \code{"total"}: Normalise the read counts to be in percent per sample.
   \item \code{"libsize"}: Normalise the read counts to adjust for gene dispersion and total read counts per sample. See \link[DESeq2]{estimateSizeFactorsForMatrix} for details.
   \item \code{"none"}: No normalisation.
   }}

\item{plot_type}{Plot type. \code{"boxplot"} or \code{"point"}. (\emph{default:} \code{"boxplot"})}

\item{point_size}{The size of points. (\emph{default:} \code{1})}

\item{row_show}{The number of variables to show, or a vector of gene/OTU names. (\emph{default:} \code{20})}

\item{order_group}{A vector to order the groups by.}

\item{order_boxgroup}{A vector to order the box-groups by.}

\item{title_newline}{Split the facet headers by newline. (\emph{default:} \code{FALSE})}

\item{title_size}{Size of text in facet headers. (\emph{default:} \code{5})}

\item{plot_flip}{(\emph{logical}) Flip the axes of the plot axis. (\emph{default:} \code{FALSE})}

\item{plot_log}{(\emph{logical}) Log10-scale the plot. (\emph{default:} \code{FALSE})}

\item{adjust_zero}{Keep values of 0 in calculations by adding this value. (\emph{default:} \code{NULL})}

\item{detailed_output}{(\emph{logical}) Return additional details or not. If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}

\item{row_labels}{(\strong{mmtravis only}) Additional variables in genedata to plot (\code{mmt}).}

\item{tax_class}{(\strong{ampvis2 only}) Converts a specific phylum to class level instead, e.g. \code{"p__Proteobacteria"}.}

\item{tax_empty}{(\strong{ampvis2 only}) How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible.
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{tax_aggregate}{(\strong{ampvis2 only}) The taxonomic level to aggregate the OTUs. (\emph{default:} \code{"Genus"})}
}
\value{
A ggplot2 object. If \code{detailed_output = TRUE} a list with a ggplot2 object and additional data.
}
\description{
Generates boxplots of genes/OTUs.
}
\details{
The plot is ordered as one would read a book, from left to right and top to bottom.
}
\examples{

\dontrun{
data("example_mmt")

vis_boxplot(example_mmt,
  group_by    = "Type",
  boxgroup_by = "Organism",
  row_show    = 10,
  row_labels  = "product",
  title_newline = T)
}
}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
