% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ordinate.R
\name{vis_ordinate}
\alias{vis_ordinate}
\title{Ordination plot}
\usage{
vis_ordinate(data, type = "", transform = "", distmeasure = "", constrain = "")
}
\arguments{
\item{obj}{(\emph{required}) A data object of class 'ampvis2' (see \href{https://github.com/MadsAlbertsen/ampvis2}{amp_load}) or class 'mmt' (see \code{\link{mt_load}}).}

\item{type}{(\emph{required}) Type of ordination method. One of:
\itemize{
   \item \code{"PCA"}: (\emph{default}) Principal Components Analysis
   \item \code{"RDA"}: Redundancy Analysis (considered the constrained version of PCA)
   \item \code{"CA"}: Correspondence Analysis
   \item \code{"CCA"}: Canonical Correspondence Analysis (considered the constrained version of CA)
   \item \code{"DCA"}: Detrended Correspondence Analysis
   \item \code{"NMDS"}: non-metric Multidimensional Scaling
   \item \code{"PCOA"} or \code{"MMDS"}: metric Multidimensional Scaling a.k.a Principal Coordinates Analysis (not to be confused with PCA)
   }
   \emph{Note that PCoA is not performed by the vegan package, but the \code{\link[ape]{pcoa}} function from the APE package.}}

\item{distmeasure}{(\emph{required for nMDS and PCoA}) Distance measure used for the distance-based ordination methods (nMDS and PCoA). Choose one of the following:
\itemize{
  \item \code{"wunifrac"}: (\strong{ampvis2 only}) Weighted generalized UniFrac distances (alpha=0.5), calculated by \code{\link[GUniFrac]{GUniFrac}}. \strong{OBS!} Requires a phylogenetic tree.
  \item \code{"unifrac"}: (\strong{ampvis2 only}) Unweighted UniFrac distances, calculated by \code{\link[GUniFrac]{GUniFrac}}. \strong{OBS!} Requires a phylogenetic tree.
  \item \code{"jsd"}: Jensen-Shannon Divergence, based on \url{http://enterotype.embl.de/enterotypes.html}.
  \item Any of the distance measures supported by \code{\link[vegan]{vegdist}}: \code{"manhattan"}, \code{"euclidean"}, \code{"canberra"}, \code{"bray"}, \code{"kulczynski"}, \code{"jaccard"}, \code{"gower"}, \code{"altGower"}, \code{"morisita"}, \code{"horn"}, \code{"mountford"}, \code{"raup"}, \code{"binomial"}, \code{"chao"}, \code{"cao"}, \code{"mahalanobis"}.
  \item or \code{"none"}. (\emph{default})
 }
You can also write your own math formula, see details in \code{\link[vegan]{vegdist}}.}

\item{transform}{(\emph{recommended}) Transforms the abundances before ordination, choose one of the following: \code{"total"}, \code{"max"}, \code{"freq"}, \code{"normalize"}, \code{"range"}, \code{"standardize"}, \code{"pa"} (presence/absense), \code{"chi.square"}, \code{"hellinger"}, \code{"log"}, or \code{"sqrt"}, see details in \code{\link[vegan]{decostand}}. Using the hellinger transformation is a good choice when performing PCA/RDA as it will produce a more ecologically meaningful result (read about the double-zero problem in Numerical Ecology). When the Hellinger transformation is used with CA/CCA it will help reducing the impact of low abundant species. When performing nMDS or PCoA (aka mMDS) it is not recommended to also use data transformation as this will obscure the chosen distance measure. (\emph{default:} \code{"hellinger"})}

\item{filter_vars}{Remove low signal variables across all samples below this threshold in percent. Setting this to 0 may drastically increase computation time. (\emph{default}: \code{0.1})}

\item{constrain}{(\emph{required for RDA and CCA}) Variable(s) in the metadata for constrained analyses (RDA and CCA). Multiple variables can be provided by a vector, fx \code{c("Year", "Temperature")}, but keep in mind that the more variables selected the more the result will be similar to unconstrained analysis.}

\item{x_axis}{Which axis from the ordination results to plot as the first axis. Have a look at the \code{$screeplot} with \code{detailed_output = TRUE} to validate axes. (\emph{default:} \code{1})}

\item{y_axis}{Which axis from the ordination results to plot as the second axis. Have a look at the \code{$screeplot} with \code{detailed_output = TRUE} to validate axes. (\emph{default:} \code{2})}

\item{sample_color_by}{Color sample points by a variable in the metadata.}

\item{sample_color_order}{Order the colors in \code{sample_color_by} by the order in a vector.}

\item{sample_shape_by}{Shape sample points by a variable in the metadata.}

\item{sample_colorframe}{Frame the sample points with a polygon by a variable in the metadata split by the variable defined by \code{sample_color_by}, or simply \code{TRUE} to frame the points colored by \code{sample_color_by}. (\emph{default:} \code{FALSE})}

\item{sample_colorframe_label}{Label by a variable in the metadata.}

\item{sample_label_by}{Label sample points by a variable in the metadata.}

\item{sample_label_size}{Sample labels text size. (\emph{default:} \code{4})}

\item{sample_label_segment_color}{Sample labels repel-segment color. (\emph{default:} \code{"black"})}

\item{sample_point_size}{Size of the sample points. (\emph{default:} \code{2})}

\item{sample_trajectory}{Make a trajectory between sample points by a variable in the metadata.}

\item{sample_trajectory_group}{Make a trajectory between sample points by the \code{sample_trajectory} argument, but within individual groups.}

\item{sample_plotly}{Enable interactive sample points so that they can be hovered to show additional information from the metadata. Provide a vector of the metadata variables to show, or \code{"all"} to display all. Click or double click the elements in the legend to hide/show parts of the data. To hide the legend use \code{plotly::layout(amp_ordinate(...), showlegend = FALSE)}, see more options at \url{https://plot.ly/r/}.}

\item{var_plot}{(\emph{logical}) Plot variable points or not. (\emph{default:} \code{FALSE})}

\item{var_nlabels}{Number of the most extreme variables labels to plot (ordered by the sum of the numerical values of the x,y coordinates. Only makes sense with PCA/RDA).}

\item{var_label_by}{Label variable points by taxonomy (\code{ampvis2}) or variable in genedata (\code{mmt}).}

\item{var_label_size}{Size of the variables text labels. (\emph{default:} \code{3})}

\item{var_label_color}{Color of the variables text labels. (\emph{default:} \code{"grey10"})}

\item{var_rescale}{(\emph{logical}) Rescale variables points or not. Basically they will be multiplied by 0.8, for visual convenience only. (\emph{default:} \code{FALSE})}

\item{var_point_size}{Size of the variables points. (\emph{default:} \code{2})}

\item{var_shape}{The shape of the variables points, fx \code{1} for hollow circles or \code{20} for dots. (\emph{default:} \code{20})}

\item{var_plotly}{(\emph{logical}) Enable interactive variables points so that they can be hovered to show complete information. (\emph{default:} \code{FALSE})}

\item{envfit_factor}{A vector of categorical environmental variables from the metadata to fit onto the ordination plot. See details in \code{\link[vegan]{envfit}}.}

\item{envfit_numeric}{A vector of numerical environmental variables from the metadata to fit arrows onto the ordination plot. The lengths of the arrows are scaled by significance. See details in \code{\link[vegan]{envfit}}.}

\item{envfit_signif_level}{The significance threshold for displaying the results of \code{envfit_factor} or \code{envfit_numeric}. (\emph{default:} \code{0.005})}

\item{envfit_textsize}{Size of the envfit text on the plot. (\emph{default:} \code{3})}

\item{envfit_textcolor}{Color of the envfit text on the plot. (\emph{default:} \code{"darkred"})}

\item{envfit_numeric_arrows_scale}{Scale the size of the numeric arrows. (\emph{default:} \code{1})}

\item{envfit_arrowcolor}{Color of the envfit arrows on the plot. (\emph{default:} \code{"darkred"})}

\item{envfit_show}{(\emph{logical}) Show the results on the plot or not. (\emph{default:} \code{TRUE})}

\item{repel_labels}{(\emph{logical}) Repel all labels to prevent cluttering of the plot. (\emph{default:} \code{TRUE})}

\item{opacity}{Opacity of all plotted points and sample_colorframe. \code{0}: invisible, \code{1}: opaque. (\emph{default:} \code{0.8})}

\item{detailed_output}{(\emph{logical}) Return additional details or not (model, scores, inputmatrix, screeplot etc). If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}

\item{tax_empty}{(\strong{ampvis2 only}) How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible.
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{...}{Pass additional arguments to the vegan ordination functions, fx the \code{\link[vegan]{rda}}, \code{\link[vegan]{cca}}, \code{\link[vegan]{metaMDS}} functions, see the documentation.}
}
\value{
A ggplot2 object. If \code{detailed_output = TRUE} a list with a ggplot2 object and additional data.
}
\description{
A wrapper around the vegan package to generate ggplot2 ordination plots suited for analysis and comparison of microbial communities. Simply choose an ordination type and a plot is returned.
}
\details{
The \code{\code{vis_ordinate}} function is primarily based on two packages; \code{\link[vegan]{vegan-package}}, which performs the actual ordination, and the \code{\link[ggplot2]{ggplot2-package}} to generate the plot. The function generates an ordination plot by the following process:
\enumerate{
  \item Various input argument checks and error messages
  \item Filtering, where low abundant variables across all samples are removed (if not \code{filter_vars = 0} is set)
  \item Data transformation (if not \code{transform = "none"} is set)
  \item Calculate distance matrix based on the chosen \code{distmeasure} if the chosen ordination method is PCoA/nMDS/DCA
  \item Perform the actual ordination and calculate the axis scores for both samples and variables
  \item Visualise the result with ggplot2 or plotly in various ways defined by the user
}
}
\references{
GUide to STatistical Analysis in Microbial Ecology (GUSTA ME): \url{https://mb3is.megx.net/gustame}

  Legendre, Pierre & Legendre, Louis (2012). Numerical Ecology. Elsevier Science. ISBN: 9780444538680

  Legendre, P., & Gallagher, E. (2001). Ecologically meaningful transformations for ordination of species data. Oecologia, 129(2), 271-280. \url{http://doi.org/10.1007/s004420100716}
}
\seealso{
\code{\link{mt_load}}
}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Thomas Yssing Michaelsen \email{tym@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
