% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_enrichKO.R
\name{mt_enrichKO}
\alias{mt_enrichKO}
\title{Perform enrichment analysis using KEGG annotation.}
\usage{
mt_enrich(mmt,GeneIDs,KOs,type)
}
\arguments{
\item{mmt}{(\emph{required}) An object of class \code{mmt}.}

\item{GeneIDs}{(\emph{required}) Character vector with gene IDs of interest.}

\item{type}{(\emph{required}) String refering to which KEGG annotation scheme to use. Choose from the following:
\itemize{
  \item \strong{pathway} - TBD.
  \item \strong{module} - TBD.
  \item \strong{reaction_class} - TBD.
  \item \strong{reaction} - TBD.
}}

\item{KOs}{(\emph{required}) String pointing to the column in mtgene containing KO identifiers.}

\item{alternative}{(\emph{optional}) String specifying the direction of the test. Choose from:
\itemize{
  \item \strong{greater} - test for enrichment in genes of interest.
  \item \strong{less} - test for depletion in genes of interest.
  \item \strong{two.sided} - perform a two-sided test for both in genes of interest.
}}
}
\value{
A data.frame
}
\description{
mt_enrichKO
}
\details{

}
\examples{

\dontrun{
# TBD.
}

}
\author{
Thomas Yssing Michaelsen \email{tym@bio.aau.dk}
}
